LPS-55710 Move
