+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/test/java/com/liferay/dynamic/data/mapping/form/field/type/date/DateDDMFormFieldTypeSettingsTest.java
@@ -12,31 +12,22 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.type.date;
+package com.liferay.dynamic.data.mapping.form.field.type.date;
 
+import com.liferay.dynamic.data.mapping.form.field.type.BaseDDMFormFieldTypeSettingsTest;
 import com.liferay.dynamic.data.mapping.model.DDMForm;
 import com.liferay.dynamic.data.mapping.model.DDMFormField;
+import com.liferay.dynamic.data.mapping.type.date.DateDDMFormFieldTypeSettings;
 import com.liferay.dynamic.data.mapping.util.DDMFormFactory;
-import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.PortalClassLoaderUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.SetUtil;
 
-import java.util.Locale;
 import java.util.Map;
-import java.util.ResourceBundle;
-import java.util.Set;
 
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import org.mockito.Mock;
-
-import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
@@ -45,14 +36,8 @@ import org.powermock.modules.junit4.PowerMockRunner;
  */
 @PrepareForTest({PortalClassLoaderUtil.class, ResourceBundleUtil.class})
 @RunWith(PowerMockRunner.class)
-public class DateDDMFormFieldTypeSettingsTest extends PowerMockito {
-
-	@Before
-	public void setUp() {
-		setUpLanguageUtil();
-		setUpPortalClassLoaderUtil();
-		setUpResourceBundleUtil();
-	}
+public class DateDDMFormFieldTypeSettingsTest
+	extends BaseDDMFormFieldTypeSettingsTest {
 
 	@Test
 	public void testCreateDateDDMFormFieldTypeSettingsDDMForm() {
@@ -70,58 +55,16 @@ public class DateDDMFormFieldTypeSettingsTest extends PowerMockito {
 			"string", predefinedValueDDMFormField.getDataType());
 		Assert.assertEquals("date", predefinedValueDDMFormField.getType());
 		Assert.assertEquals(true, predefinedValueDDMFormField.isLocalizable());
+
+		DDMFormField validationDDMFormField = ddmFormFieldsMap.get(
+			"validation");
+
+		Assert.assertNotNull(validationDDMFormField);
+		Assert.assertEquals(
+			"ddm-validation", validationDDMFormField.getDataType());
+		Assert.assertEquals("validation", validationDDMFormField.getType());
+		Assert.assertEquals(
+			"false", validationDDMFormField.getVisibilityExpression());
 	}
-
-	protected void setUpLanguageUtil() {
-		Set<Locale> availableLocales = SetUtil.fromArray(
-			new Locale[] {LocaleUtil.US});
-
-		when(
-			language.getAvailableLocales()
-		).thenReturn(
-			availableLocales
-		);
-
-		LanguageUtil languageUtil = new LanguageUtil();
-
-		languageUtil.setLanguage(language);
-	}
-
-	protected void setUpPortalClassLoaderUtil() {
-		mockStatic(PortalClassLoaderUtil.class);
-
-		when(
-			PortalClassLoaderUtil.getClassLoader()
-		).thenReturn(
-			_classLoader
-		);
-	}
-
-	protected void setUpResourceBundleUtil() {
-		mockStatic(ResourceBundleUtil.class);
-
-		when(
-			ResourceBundleUtil.getBundle(
-				"content.Language", LocaleUtil.BRAZIL, _classLoader)
-		).thenReturn(
-			_resourceBundle
-		);
-
-		when(
-			ResourceBundleUtil.getBundle(
-				"content.Language", LocaleUtil.US, _classLoader)
-		).thenReturn(
-			_resourceBundle
-		);
-	}
-
-	@Mock
-	protected Language language;
-
-	@Mock
-	private ClassLoader _classLoader;
-
-	@Mock
-	private ResourceBundle _resourceBundle;
 
 }
