+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/test/java/com/liferay/dynamic/data/mapping/form/field/type/radio/RadioDDMFormFieldValueAccessorTest.java
@@ -12,13 +12,13 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.type.select;
+package com.liferay.dynamic.data.mapping.form.field.type.radio;
 
 import com.liferay.dynamic.data.mapping.model.UnlocalizedValue;
 import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
 import com.liferay.dynamic.data.mapping.test.util.DDMFormValuesTestUtil;
+import com.liferay.dynamic.data.mapping.type.radio.RadioDDMFormFieldValueAccessor;
 import com.liferay.portal.json.JSONFactoryImpl;
-import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.util.LocaleUtil;
 
@@ -29,45 +29,32 @@ import org.junit.Test;
 /**
  * @author Renato Rego
  */
-public class SelectDDMFormFieldValueAccessorTest {
+public class RadioDDMFormFieldValueAccessorTest {
 
 	@Before
 	public void setUp() {
 		setUpJSONFactoryUtil();
 	}
 
-	public void setUpJSONFactoryUtil() {
+	@Test
+	public void testGetRadioValue() {
+		DDMFormFieldValue ddmFormFieldValue =
+			DDMFormValuesTestUtil.createDDMFormFieldValue(
+				"Radio", new UnlocalizedValue("[\"value 1\"]"));
+
+		RadioDDMFormFieldValueAccessor radioDDMFormFieldValueAccessor =
+			new RadioDDMFormFieldValueAccessor();
+
+		Assert.assertEquals(
+			"value 1",
+			radioDDMFormFieldValueAccessor.getValue(
+				ddmFormFieldValue, LocaleUtil.US));
+	}
+
+	protected void setUpJSONFactoryUtil() {
 		JSONFactoryUtil jsonFactoryUtil = new JSONFactoryUtil();
 
 		jsonFactoryUtil.setJSONFactory(new JSONFactoryImpl());
-	}
-
-	@Test
-	public void testGetSelectValue() throws Exception {
-		JSONArray expectedJSONArray = createJSONArray("value 1");
-
-		DDMFormFieldValue ddmFormFieldValue =
-			DDMFormValuesTestUtil.createDDMFormFieldValue(
-				"Select", new UnlocalizedValue(expectedJSONArray.toString()));
-
-		SelectDDMFormFieldValueAccessor selectDDMFormFieldValueAccessor =
-			new SelectDDMFormFieldValueAccessor();
-
-		JSONArray actualJSONArray = selectDDMFormFieldValueAccessor.getValue(
-			ddmFormFieldValue, LocaleUtil.US);
-
-		Assert.assertEquals(
-			expectedJSONArray.toString(), actualJSONArray.toString());
-	}
-
-	protected JSONArray createJSONArray(String... strings) {
-		JSONArray jsonArray = JSONFactoryUtil.createJSONArray();
-
-		for (String string : strings) {
-			jsonArray.put(string);
-		}
-
-		return jsonArray;
 	}
 
 }
