+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/test/java/com/liferay/dynamic/data/mapping/form/field/type/radio/RadioDDMFormFieldValueRendererTest.java
@@ -12,35 +12,34 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.type.date;
+package com.liferay.dynamic.data.mapping.form.field.type.radio;
 
 import com.liferay.dynamic.data.mapping.model.DDMForm;
 import com.liferay.dynamic.data.mapping.model.DDMFormField;
+import com.liferay.dynamic.data.mapping.model.DDMFormFieldOptions;
 import com.liferay.dynamic.data.mapping.model.UnlocalizedValue;
 import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
 import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
 import com.liferay.dynamic.data.mapping.test.util.DDMFormTestUtil;
 import com.liferay.dynamic.data.mapping.test.util.DDMFormValuesTestUtil;
-import com.liferay.portal.kernel.util.DateFormatFactoryUtil;
-import com.liferay.portal.kernel.util.FastDateFormatFactoryUtil;
+import com.liferay.dynamic.data.mapping.type.radio.RadioDDMFormFieldValueAccessor;
+import com.liferay.dynamic.data.mapping.type.radio.RadioDDMFormFieldValueRenderer;
+import com.liferay.portal.json.JSONFactoryImpl;
+import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.util.DateFormatFactoryImpl;
-import com.liferay.portal.util.FastDateFormatFactoryImpl;
 
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
 /**
- * @author Bruno Basto
+ * @author Renato Rego
  */
-public class DateDDMFormFieldValueRendererTest {
+public class RadioDDMFormFieldValueRendererTest {
 
 	@Before
 	public void setUp() {
-		setUpDateFormatFactoryUtil();
-		setUpFastDateFormatFactoryUtil();
+		setUpJSONFactoryUtil();
 	}
 
 	@Test
@@ -48,7 +47,15 @@ public class DateDDMFormFieldValueRendererTest {
 		DDMForm ddmForm = DDMFormTestUtil.createDDMForm();
 
 		DDMFormField ddmFormField = DDMFormTestUtil.createDDMFormField(
-			"birthday", "Birthday", "date", "string", false, false, false);
+			"Radio", "Radio", "radio", "string", false, false, false);
+
+		DDMFormFieldOptions ddmFormFieldOptions =
+			ddmFormField.getDDMFormFieldOptions();
+
+		ddmFormFieldOptions.addOptionLabel(
+			"value 1", LocaleUtil.US, "option 1");
+		ddmFormFieldOptions.addOptionLabel(
+			"value 2", LocaleUtil.US, "option 2");
 
 		ddmForm.addDDMFormField(ddmFormField);
 
@@ -57,43 +64,35 @@ public class DateDDMFormFieldValueRendererTest {
 
 		DDMFormFieldValue ddmFormFieldValue =
 			DDMFormValuesTestUtil.createDDMFormFieldValue(
-				"birthday", new UnlocalizedValue("2015-01-25"));
+				"Radio", new UnlocalizedValue("[\"value 1\"]"));
 
 		ddmFormValues.addDDMFormFieldValue(ddmFormFieldValue);
 
-		DateDDMFormFieldValueRenderer dateDDMFormFieldValueRenderer =
-			new DateDDMFormFieldValueRenderer();
+		RadioDDMFormFieldValueRenderer radioDDMFormFieldValueRenderer =
+			createRadioDDMFormFieldValueRenderer();
 
 		Assert.assertEquals(
-			"1/25/15",
-			dateDDMFormFieldValueRenderer.render(
-				ddmFormFieldValue, LocaleUtil.US));
-		Assert.assertEquals(
-			"25/01/15",
-			dateDDMFormFieldValueRenderer.render(
-				ddmFormFieldValue, LocaleUtil.BRAZIL));
-
-		ddmFormFieldValue.setValue(new UnlocalizedValue(""));
-
-		Assert.assertEquals(
-			StringPool.BLANK,
-			dateDDMFormFieldValueRenderer.render(
+			"option 1",
+			radioDDMFormFieldValueRenderer.render(
 				ddmFormFieldValue, LocaleUtil.US));
 	}
 
-	protected void setUpDateFormatFactoryUtil() {
-		DateFormatFactoryUtil dateFormatFactoryUtil =
-			new DateFormatFactoryUtil();
+	protected RadioDDMFormFieldValueRenderer
+		createRadioDDMFormFieldValueRenderer() {
 
-		dateFormatFactoryUtil.setDateFormatFactory(new DateFormatFactoryImpl());
+		RadioDDMFormFieldValueRenderer radioDDMFormFieldValueRenderer =
+			new RadioDDMFormFieldValueRenderer();
+
+		radioDDMFormFieldValueRenderer.setRadioDDMFormFieldValueAccessor(
+			new RadioDDMFormFieldValueAccessor());
+
+		return radioDDMFormFieldValueRenderer;
 	}
 
-	protected void setUpFastDateFormatFactoryUtil() {
-		FastDateFormatFactoryUtil fastDateFormatFactoryUtil =
-			new FastDateFormatFactoryUtil();
+	protected void setUpJSONFactoryUtil() {
+		JSONFactoryUtil jsonFactoryUtil = new JSONFactoryUtil();
 
-		fastDateFormatFactoryUtil.setFastDateFormatFactory(
-			new FastDateFormatFactoryImpl());
+		jsonFactoryUtil.setJSONFactory(new JSONFactoryImpl());
 	}
 
 }
